Add files via upload
Car Parking - Application of Stacks and Queues

It is a car parking programming model where 5 cars will be present in a normal queue at a time. When more cars are ready to park, they will be directed to a waiting queue.
When a car is removed from the normal queue, automatically one car is dequeued from waiting queue and inserted into actual queue through the rear end. If a car is to be removed from the normal queue, all the cars before it will be pushed to a stack and then removed.